+++ app/controllers/steps_controller.rb
@@ -10,6 +10,11 @@ class StepsController < ApplicationController
     @step_presenter = StepPresenter.new(@step)
 
     @answer = AnswerFactory.new(step: @step).call
+    @back_url = if !parent_task || parent_task.has_single_visible_step?
+      journey_path(@journey, anchor: @step.id)
+    else
+      journey_task_path(@journey, parent_task)
+    end
 
     render @step.contentful_type, locals: {layout: "steps/new_form_wrapper"}
   end
@@ -21,11 +26,20 @@ class StepsController < ApplicationController
     @step_presenter = StepPresenter.new(@step)
 
     @answer = @step.answer
+    @back_url = if !parent_task || parent_task.has_single_visible_step?
+      journey_path(@journey, anchor: @step.id)
+    else
+      journey_task_path(@journey, parent_task)
+    end
 
     render "steps/#{@step.contentful_type}", locals: {layout: "steps/edit_form_wrapper"}
   end
 
   private
+
+  def parent_task
+    @step.task
+  end
 
   def journey_id
     params[:journey_id]